requirements.txt
fastapi
uvicorn
google.generativeai
PyPDF2==3.0.1
python-dotenv
python-multipart